added UtilsTest.java removed clamplooped int version

+++ src/java/penoplatinum/simulator/Simulator.java
@@ -111,7 +111,7 @@ public class Simulator {
    */
   int findHitDistance(int angle, int left, int top, double x, double y) {
     // Force angles between 0 and 360 !!!
-    angle = penoplatinum.util.Utils.ClampLooped(angle, 0, 360);
+    angle = (int) penoplatinum.util.Utils.ClampLooped(angle, 0, 360);
     //if (angle < 0 || angle > 360) throw new IllegalArgumentException();
 
     // determine the point on the (virtual) wall on the current tile, where
